fix: edit wallets lesson (#4)
* Capitalize Devnet

The hyperlink is optional, but does provide an oppty to learn where to 
check it's status?

* Maybe it's changed?

Using the Brave browser, my Phantom extension v22.4.8 has the settings 
gear icon on the right 🤷🏻‍♂️

* awk order

now it goes from wallets > hardware > software > software type & support

* flow from general to specific example

* awk tab indents

not sute what causes this, but the \t spacing made the indents render up 
all f'd up and out of whack, so I just copied the indents from lines 
which "lined up" proper like

* awk

Given the text placement after the image, so it reads like a comment 
upon or clarification of what has been achived in the example

* consistent tabs for WalletContextProvider.tsx

* wallets formatting & component specifier

The formatting is to be consistent.  The component naming 
(WalletContextProvider) is prolly overly on the nose - I just dislike 
the redundnt/ambiguous React terminology of "import these componenets.  
Add them to the component."

* awk

* grouping the import comments in one paragraph

* No need to repeat all the import statements

Tightening up section 3 of the demo.  I get why you made the endpoint / 
wallet a section, then added the require statement for styling, but this 
edit is a leaner wrap :D

* consistent tab style

* remove extra newline from codeblock

* Pretty sure we still need to import PingButton here?

* Noting the update from "Connect to "Select Wallet"

* is awk calling it the connect button

* consistency

* verb tense

* indentation

* so six-y!